Create service method signatures for plant routes.
I created the service method signatures for the plant routes and wrote
the code for the get_all_user_plants method.

Also added todo to remove the get_user service method and route as this
would provide a security risk in production but could still be
convenient for future testing purposes in development.